move logout to Account page
