+++ src/components/Account.vue
@@ -2,12 +2,19 @@
   <main class="account">
       <h1>Account</h1>
       <Auth v-if="!this.$store.state.user.authenticated"></Auth>
-      <p v-else>Logged in as {{ this.$store.state.user.email }}</p>
+      <template v-else>
+       <p>Logged in as {{ this.$store.state.user.email }}</p>
+        <el-button type="info" @click="googleLogoutHandler">
+          Logout
+        </el-button>
+      </template>
   </main>
 </template>
 
 <script>
 import Auth from '@/components/Auth'
+import { googleLogout } from '@/api'
+
 
 export default {
   name: 'Account',
@@ -16,6 +23,17 @@ export default {
     return {
       authenticated: this.$store.state.user.authenticated,
       email: this.$store.state.user.email
+    }
+  },
+  methods: {
+    async googleLogoutHandler () {
+      try {
+        await googleLogout()
+        this.$store.commit('logout')
+        this.$router.push('/')
+      } catch (e) {
+        console.error(e)
+      }
     }
   }
 }
