+++ src/components/NavbarDesktop.vue
@@ -18,12 +18,6 @@
             </el-button>
           </router-link>
         </el-menu-item>
-        <!-- move below to inside account page -->
-        <el-menu-item index="6" v-if="this.$store.state.user.authenticated">
-            <el-button type="info" @click="googleLogoutHandler">
-              Logout
-            </el-button>
-        </el-menu-item>
     </el-menu>
   </el-header>
 </template>
@@ -31,7 +25,7 @@
 <script>
 export default {
   name: 'NavbarDesktop',
-  props: ['googleLogoutHandler', 'navigation', 'authenticated', 'email'],
+  props: [ 'navigation', 'authenticated', 'email'],
   data () {
     return {
       activeIndex: '0'
